small changes around projects. clean up
