+++ src/projects-v2/GroceryLists/src/index.js
@@ -1,3 +1,6 @@
+import files from './files'
+
+
 import _ from 'lodash'
