+++ src/projects-v2/Nutritions/src/index.js
@@ -1,0 +1,1 @@
+import files from './files'
