+++ src/projects-v2/UnitConvertion/src/index.js
@@ -1,1 +1,1 @@
-import files from './../files.js'
+import files from './files'
