+++ src/projects-v2/showcase-grocery-lists/src/index.js
@@ -1,0 +1,1 @@
+import files from './files'
