+++ src/projects-v2/showcase-recipes/src/files.js
@@ -1,0 +1,11 @@
+import {
+  ingredients3,
+  recipes,
+  departments
+} from '@groceristar/sd-wrapper'
+
+export {
+  ingredients3,
+  recipes,
+  departments
+}
